Small refactoring + working timer

+++ ChessApp.Server/Models/Move.cs
@@ -5,6 +5,6 @@
         public required int Row { get; set; }
         public required int Column { get; set; }
         public required Piece Piece { get; set; }
-        public required int TimeLeft { get; set; }
+        public required float TimeLeft { get; set; }
     }
 }
